test(linter/plugins): move test group definitions into separate files (#18503)
Refactor conformance tests. Move group definitions into their own files, to make room for adding more repos to the tests.
