+++ apps/oxlint/conformance/src/groups/index.ts
@@ -1,0 +1,6 @@
+import type { TestGroup } from "../index.ts";
+
+import eslint from "./eslint.ts";
+import reactHooks from "./react_hooks.ts";
+
+export const TEST_GROUPS: TestGroup[] = [eslint, reactHooks];
